+++ src/Akeneo/Pim/Enrichment/Component/Error/DomainErrorInterface.php
@@ -8,6 +8,6 @@ namespace Akeneo\Pim\Enrichment\Component\Error;
  * @copyright 2020 Akeneo SAS (http://www.akeneo.com)
  * @license http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-interface DomainErrorInterface
+interface DomainErrorInterface extends \Throwable
 {
 }